+++ app/Http/Controllers/PositionController.php
@@ -6,9 +6,12 @@ use App\Models\Position;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Inertia\Inertia;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class PositionController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index(Request $request)
     {
         $query = Position::query();
@@ -20,20 +23,27 @@ class PositionController extends Controller
 
         return Inertia::render('Positions/Index', [
             'positions' => $query->paginate(25)
-                               ->withQueryString()
+                               ->withQueryString(),
+            'can' => [
+                'create' => $request->user()->can('create', Position::class),
+                'edit' => $request->user()->can('update', Position::class),
+                'delete' => $request->user()->can('delete', Position::class),
+            ]
         ]);
     }
 
     public function create()
     {
+        $this->authorize('create', Position::class);
         return Inertia::render('Positions/Create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Position::class);
+
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:255',
-            'description' => 'nullable|string',
         ]);
 
         if ($validator->fails()) {
@@ -48,6 +58,7 @@ class PositionController extends Controller
 
     public function edit(Position $position)
     {
+        $this->authorize('update', $position);
         return Inertia::render('Positions/Edit', [
             'position' => $position
         ]);
@@ -55,9 +66,10 @@ class PositionController extends Controller
 
     public function update(Request $request, Position $position)
     {
+        $this->authorize('update', $position);
+
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:255',
-            'description' => 'nullable|string',
         ]);
 
         if ($validator->fails()) {
@@ -72,6 +84,8 @@ class PositionController extends Controller
 
     public function destroy(Position $position)
     {
+        $this->authorize('delete', $position);
+        
         $position->delete();
 
         return redirect()->route('positions.index')
